pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.0",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pyyaml>=6.0.2",
]
